Internal mod, shadow fabric-loader api, and initial entrypoints implementation
A separate standard Forge mod is now built and included within the main jar, and loaded by the ModLocator at runtime, as Forge does not load mod data directly from jars containing ModLocators.
fabric-loader's API is now shaded into the main jar as well; currently this is just for the ModInitializer interfaces, and no API is actually implemented.
Added a basic entrypoints implementation; this currently only handles the client-side, and only handles cases where the entrypoint is a class (not a method or field reference).