add a dummy subproject to allow launching in dev environments
Forge allows for loading mod locators as dependencies in development environments, but does not support loading them directly in their own project. As such, we must make a dummy mod that depends on Steelwool in order to load it in a development environment.
(To do: find a workaround that allows us to avoid doing this)