Temporary hack for slightly better compatibility until we rewrite the remapper
There are some methods that are not obfuscated, and thus have no intermediary mapping, but still have a tsrg mapping.
Because they're not obfuscated, their names are also not unique. Thus, we special-case these method names to also consider the owner of the method.
This is a hack because we only consider the direct owner; for subclasses this will break, which is an unavoidable problem until we rewrite the remapper.